Rename package from wagtailannotatedimage -> wagtail_annotations
