+++ wagtail_annotations/frontend/js/annotated-image-handler.js
