+++ wagtail_annotations/static/annotated-image/annotated-image.css
