+++ wagtail_annotations/static/annotated-image/jquery.annotate.js
