+++ wagtail_annotations/templates/annotated_image.html
