+++ wagtail_annotations/templates/annotated_image.js
